refactor: organize dto packages by responsibility
