--- src/main/java/com/financebot/ai/dto/AiParseResponse.java
@@ -1,4 +1,0 @@
-package com.financebot.ai.dto;
-
-public class AiParseResponse {
-}
