--- src/main/java/com/financebot/ai/dto/ConfidenceResult.java
@@ -1,4 +1,0 @@
-package com.financebot.ai.dto;
-
-public class ConfidenceResult {
-}
