+++ src/main/java/com/financebot/ai/dto/request/AiParseRequest.java
@@ -1,0 +1,4 @@
+package com.financebot.ai.dto.request;
+
+public class AiParseRequest {
+}
