+++ src/main/java/com/financebot/ai/dto/response/AiParseResponse.java
@@ -1,0 +1,4 @@
+package com.financebot.ai.dto.response;
+
+public class AiParseResponse {
+}
