+++ src/main/java/com/financebot/ai/dto/response/ConfidenceResult.java
@@ -1,0 +1,4 @@
+package com.financebot.ai.dto.response;
+
+public class ConfidenceResult {
+}
