+++ src/main/java/com/financebot/auth/controller/AuthController.java
@@ -1,8 +1,8 @@
 package com.financebot.auth.controller;
 
-import com.financebot.auth.dto.AuthResponse;
-import com.financebot.auth.dto.LoginRequest;
-import com.financebot.auth.dto.RegisterRequest;
+import com.financebot.auth.dto.response.AuthResponse;
+import com.financebot.auth.dto.request.LoginRequest;
+import com.financebot.auth.dto.request.RegisterRequest;
 import com.financebot.auth.service.AuthService;
 import jakarta.validation.Valid;
 import org.springframework.http.HttpStatus;
