+++ src/main/java/com/financebot/auth/dto/response/AuthResponse.java
@@ -1,4 +1,4 @@
-package com.financebot.auth.dto;
+package com.financebot.auth.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
