+++ src/main/java/com/financebot/category/dto/request/CreateCategoryRequest.java
@@ -1,11 +1,11 @@
-package com.financebot.category.dto;
+package com.financebot.category.dto.request;
 
 import com.financebot.category.domain.CategoryType;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
 
-public record UpdateCategoryRequest(
+public record CreateCategoryRequest(
 
         @NotBlank(message = "Name is required")
         @Size(max = 100, message = "Name must have at most 100 characters")
