+++ src/main/java/com/financebot/category/service/CategoryService.java
@@ -2,9 +2,9 @@ package com.financebot.category.service;
 
 import com.financebot.category.domain.Category;
 import com.financebot.category.domain.CategoryType;
-import com.financebot.category.dto.CategoryResponse;
-import com.financebot.category.dto.CreateCategoryRequest;
-import com.financebot.category.dto.UpdateCategoryRequest;
+import com.financebot.category.dto.response.CategoryResponse;
+import com.financebot.category.dto.request.CreateCategoryRequest;
+import com.financebot.category.dto.request.UpdateCategoryRequest;
 import com.financebot.category.mapper.CategoryMapper;
 import com.financebot.category.repository.CategoryRepository;
 import com.financebot.user.domain.User;
